add prelude, add plugin, move module
