+++ src/prelude.rs
@@ -1,0 +1,2 @@
+pub use crate::TimerPlugin;
+pub use crate::despawn_after_timer;
